// backend/infrastructure/cdk.out/asset.0bdef5a2cc2d57e707ea7073217ee64c7352ff5dafd182474b36fd78b219df0c/types/api.ts
export interface PaginatedResponse<T> {
  items: T[];
  nextToken?: string;
  total?: number;
}
// ...
export interface AccountSummary {
  account_id: string;
  username: string;
  handle: string;
  profile_image?: string;
  account_type: AccountType;
  is_private: boolean;
}
// ...
export interface MessageDetail {
  message_id: string;
  conversation_id: string;
  content: string;
  sender: AccountSummary;
  created_at: number;
  is_read: boolean;
}

export interface GetMessagesResponse extends PaginatedResponse<MessageDetail> {}